data validation

+++ Controllers/CommentController.cs
@@ -19,23 +19,33 @@ namespace api.Controllers
     {
         private readonly ICommentRepository _commentRepository;
 
-        public CommentController(ICommentRepository commentRepository)
+        private readonly IStockRepository _stockRepository;
+        public CommentController(ICommentRepository commentRepository, IStockRepository stockRepository)
         {
             _commentRepository = commentRepository;
+            _stockRepository = stockRepository;
         }
 
         // GET: api/Comment
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(x => x.ToCommentDto());
             return Ok(commentDto);
         }
         // GET: api/Comment/5
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.GetByIdAsync(id);
 
             if (comment == null)
@@ -48,39 +58,55 @@ namespace api.Controllers
 
         // PUT: api/Comment/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] Comment commentModel)
+        [HttpPut("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateCommentRequestDto)
         {
-            var comment = await _commentRepository.UpdateAsync(id, commentModel);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            var comment = await _commentRepository.UpdateAsync(id, updateCommentRequestDto.ToCommentFromUpdateDTO());
             if (comment == null)
             {
-                return NotFound();
+                return NotFound("Comment not found!");
             }
             return Ok(comment.ToCommentDto());
         }
 
         // POST: api/Comment
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<IActionResult> Create([FromBody] CreateCommentRequestDto createCommentRequestDto)
+        [HttpPost("{stockId:int}")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentRequestDto createCommentRequestDto)
         {
-            var commentModel = createCommentRequestDto.ToCommentFromCreateDTO();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            if (!await _stockRepository.StockExists(stockId))
+            {
+                return BadRequest("Stock does not exist!");
+            }
+            var commentModel = createCommentRequestDto.ToCommentFromCreateDTO(stockId);
             await _commentRepository.CreateAsync(commentModel);
 
-            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel);
+            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
         // DELETE: api/Comment/5
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteComment([FromRoute]int id)
+        [HttpDelete("{id:int}")]
+        public async Task<IActionResult> DeleteComment([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.DeleteAsync(id);
             if (comment == null)
             {
-                return NotFound();
+                return NotFound("Comment does not exist!");
             }
 
-            return NoContent();
+            return Ok(comment);
         }
     }
 }
